requirements.txt
annotated-types==0.7.0
anyio==4.9.0
art==6.5
asgiref==3.9.0
beautifulsoup4==4.12.3
CacheControl==0.14.3
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==45.0.5
dj-database-url==3.0.1
Django==4.2.23
exceptiongroup==1.3.0
firebase_admin==7.0.0
gcloud==0.18.3
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jwcrypto==1.5.6
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.1
oauth2client==4.1.3
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
Pyrebase4==4.8.0
python-dotenv==1.1.1
python-jwt==4.1.0
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==1.26.20
uvicorn==0.35.0
whitenoise==6.9.0
